requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
djangorestframework==3.13.1
Faker==13.3.5
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
